Create C# - Converting Data Types
ATTEMPT 1 - To start, below the Console.Write() statement, create an int variable named faveNumber and set it equal to Console.ReadLine().
( error will be received )
ATTEMPT 2 - Try explicitly casting the value of faveNumber as an int 
(error will be received ) 
ATTEMPT 3 - Look at this article on converting strings to int. It lists a few of the methods in the Convert class, including: Convert.ToInt32(). This method takes a string and outputs an integer.